[FIVE-301] GP: Se modificó el tipo de corrigió el servicio AwsArtifactsProviderService para tener en cuenta que el tipo de dato de PricePerUnit de PricingDimension es decimal en lugar de float

+++ FiveRockingFingers/FRF.Core/Services/AwsArtifactsProviderService.cs
@@ -223,7 +223,7 @@ namespace FRF.Core.Services
                 RateCode = (string)termPriceDimension.Value.SelectToken("rateCode"),
                 BeginRange = beginRange,
                 Currency = termPriceDimension.Value.SelectToken("pricePerUnit").ToObject<JObject>().Properties().ElementAt(0).Name,
-                PricePerUnit = (float)termPriceDimension.Value.SelectToken("pricePerUnit").ToObject<JObject>().Properties().ElementAt(0).Value,
+                PricePerUnit = (decimal)termPriceDimension.Value.SelectToken("pricePerUnit").ToObject<JObject>().Properties().ElementAt(0).Value,
             };
 
             return pricingDetails;
